// general_examples/src/types/types.d.ts
/*-
 * #%L
 * vmware_aria_orchestrator_examples
 * %%
 * Copyright (C) 2025 TODO: Enter Organization name
 * %%
 * TODO: Define header text
 * #L%
 */
export interface SampleInterface {
  foo: string;
  bar: string;
}

export type YesNo = "Yes" | "No";

type AttributeMap = {
  powerMin?: number;
  powerMax?: number;
  domainName?: string;
};